requirements.txt
flask>=2.3.3
rembg[cpu]>=2.0.50,<3.0.0
Pillow>=10.0.0
flask-wtf>=1.2.1
flask-limiter>=3.5.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
python-json-logger>=2.0.7
email-validator>=2.1.0.post1
requests>=2.31.0
flask-cors>=4.0.0
